[AB#40224] refactor: rename encoding to video service
