+++ services/entitlement/service/.env.template
@@ -32,7 +32,7 @@ GEOLITE2_LICENSE_KEY=
 
 # These variables are optional and are only needed for local development
 DEV_USER_SERVICE_MANAGEMENT_BASE_URL=https://user.service.eu.axinom.net
-DEV_ENCODING_SERVICE_BASE_URL=https://encoding.service.eu.axinom.net
+DEV_VIDEO_SERVICE_BASE_URL=https://encoding.service.eu.axinom.net
 DEV_APPLICATION_NAME="DEV Application (entitlement-service)"
 DEV_END_USER_ID=d68f0268-2703-49ef-9239-7396073fe77a
