+++ services/media/service/.env.template
@@ -13,7 +13,7 @@ DATABASE_LOGIN=media_service_login
 DATABASE_LOGIN_PASSWORD=media_service_login_pwd
 DATABASE_GQL_ROLE=media_service_gql_role
 
-ENCODING_SERVICE_BASE_URL=https://encoding.service.eu.axinom.net
+VIDEO_SERVICE_BASE_URL=https://encoding.service.eu.axinom.net
 IMAGE_SERVICE_BASE_URL=https://image.service.eu.axinom.net
 
 SERVICE_ACCOUNT_CLIENT_ID=####
