+++ services/media/service/src/domains/movies/handlers/publishing-movie-processor.db.spec.ts
@@ -602,7 +602,7 @@ describe('publishingMovieProcessor', () => {
                 title: 'a',
                 output_format: 'DASH',
                 length_in_seconds: 1,
-                audio_languages: ['en'], // Trusting encoding service
+                audio_languages: ['en'], // Trusting Video Service
                 subtitle_languages: [],
                 caption_languages: [],
                 dash_manifest:
