+++ services/media/service/src/domains/tvshows/handlers/publishing-episode-processor.db.spec.ts
@@ -606,7 +606,7 @@ describe('publishingEpisodeProcessor', () => {
                 title: 'a',
                 output_format: 'DASH',
                 length_in_seconds: 1,
-                audio_languages: ['en'], // Trusting encoding service
+                audio_languages: ['en'], // Trusting Video Service
                 subtitle_languages: [],
                 caption_languages: [],
                 dash_manifest:
