+++ services/media/service/src/domains/tvshows/handlers/publishing-season-processor.ts
@@ -61,7 +61,7 @@ const seasonDataAggregator: SnapshotDataAggregator = async (
 
   const { result: videos, validation: videosValidation } =
     await getVideosMetadata(
-      config.encodingServiceBaseUrl,
+      config.videoServiceBaseUrl,
       authToken,
       null,
       season.trailers,
